// api/src/model/Planta.ts
export interface Planta {
    id?: number;
    nome_comum?: string;
    nome_personalizado?: string;
    nome_cientifico?: string;
    proxima_adubagem?: string;
    ciclo?: string;
    regagem?: string;
    banho_sol?: string;
    user_id?: number;
    planta_base_id?: number;
    imagem_url?: string;
  }